Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCDataChannel-id-expected.txt


FAIL DTLS client uses odd data channel IDs assert_equals: id expected 1 but got 0
FAIL DTLS server uses even data channel IDs assert_not_equals: got disallowed value null
